third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/importNode-expected.txt

Test that template contents are correctly cloned when the template element is imported.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS template.content.childNodes.length is 1
PASS imported.content.childNodes.length is 1
PASS imported.outerHTML is template.outerHTML
PASS imported.content.firstChild is not template.content.firstChild
PASS imported.content is not template.content
PASS imported.firstChild.tagName is "DIV"
PASS imported.firstChild is not div
PASS successfullyParsed is true

TEST COMPLETE
